Include a __main__ method so the module can run from the CLI.
When the module is run directly from the CLI - ie,

> python printpack.py

its __name__ variable is set to __main__ (whereas in other circumstances, it's
'printpack'). This code at the end of the file will automatically execute
print_forms(), which seems to work just fine when taking its inputs from the
Windows command line (although it seems to screw up MinGW).